agent/requirements.txt
openai>=1.60.0
python-dotenv>=1.0.0
mcp>=1.2.0
websockets>=13.0
sounddevice>=0.5.0
numpy>=1.26.0
